feat(options_ui): allow user to input sites
Implement logic and basic UI to allow users to add sites they want to
block to the dynamic ruleset using the declarativeNetRequest api to get
and set rules.

Remove storage.local as it offers no benefits over the dynamic ruleset
for the extension in terms of persistant storage and access of dynamic
ruleset.

Remove background.js from manifest due to script loading error,
background.js is currently not needed.