feat(shield): Improve blocking
Add sub_frame option to resourceTypes property for rules in the dynamic
rulset to block redirect links to blocked sites.

Change the action type property for rules in the dynamic ruleset from
"block" to "redirect" to ensure reliable blocking where the site is
being served from the browser cache, leading to a lack of network
requests for the dynamic ruleset to redirect.

Add shield.html to redirect blocked urls to custom page served by the
extension.

Update vite build logic to copy shield.html to the dist folder to make
it accessible to manifest and redirect logic.